@@ -5,6 +5,8 @@ import android.graphics.drawable.Drawable;
 import android.graphics.drawable.StateListDrawable;
 import android.widget.ImageView;
 
+import com.wei.android.lib.colorview.utils.Constant;
+
 /**
  * Created by UCCMAWEI on 2017/11/17.
  */
@@ -12,90 +14,272 @@ import android.widget.ImageView;
 public class ColorImageViewHelper extends ColorViewHelper<ImageView> {
 
     // 图片
-    private Drawable mSrcDrawableNormal;
-    private Drawable mSrcDrawablePressed;
-    private Drawable mSrcDrawableSelected;
-    private Drawable mSrcDrawableChecked;
-    private Drawable mSrcDrawableUnable;
+    private Drawable mSrcNormal;
+    private Drawable mSrcPressed;
+    private Drawable mSrcSelected;
+    private Drawable mSrcChecked;
+    private Drawable mSrcUnable;
 
-    public ColorImageViewHelper(
-            ImageView view,
-            TypedArray typedArray,
-            int srcDrawableNormal,
-            int srcDrawablePressed,
-            int srcDrawableSelected,
-            int srcDrawableChecked,
-            int srcDrawableUnable,
-            int bgColorNormal,
-            int bgColorPressed,
-            int bgColorSelected,
-            int bgColorChecked,
-            int bgColorUnable,
-            int bgDrawableNormal,
-            int bgDrawablePressed,
-            int bgDrawableSelected,
-            int bgDrawableChecked,
-            int bgDrawableUnable,
-            int bdColorNormal,
-            int bdColorPressed,
-            int bdColorSelected,
-            int bdColorChecked,
-            int bdColorUnable,
-            int bdWidth,
-            int bdDashGap,
-            int bdDashWidth,
-            int cr,
-            int crTopLeft,
-            int crTopRight,
-            int crBottomLeft,
-            int crBottomRight
-    ) {
-        super(
-                view, typedArray,
-                bgColorNormal,
-                bgColorPressed,
-                bgColorSelected,
-                bgColorChecked,
-                bgColorUnable,
-                bgDrawableNormal,
-                bgDrawablePressed,
-                bgDrawableSelected,
-                bgDrawableChecked,
-                bgDrawableUnable,
-                bdColorNormal,
-                bdColorPressed,
-                bdColorSelected,
-                bdColorChecked,
-                bdColorUnable,
-                bdWidth,
-                bdDashGap,
-                bdDashWidth,
-                cr,
-                crTopLeft,
-                crTopRight,
-                crBottomLeft,
-                crBottomRight);
+    public ColorImageViewHelper(ImageView imageView,
+                                TypedArray typedArray,
 
-        mSrcDrawableNormal = typedArray.getDrawable(srcDrawableNormal);
-        mSrcDrawablePressed = typedArray.getDrawable(srcDrawablePressed);
-        mSrcDrawableSelected = typedArray.getDrawable(srcDrawableSelected);
-        mSrcDrawableChecked = typedArray.getDrawable(srcDrawableChecked);
-        mSrcDrawableUnable = typedArray.getDrawable(srcDrawableUnable);
+                                int srcNormal,
+                                int srcPressed,
+                                int srcSelected,
+                                int srcChecked,
+                                int srcUnable,
 
-        if (mSrcDrawablePressed == null && mSrcDrawableNormal != null) {
-            mSrcDrawablePressed = typedArray.getDrawable(srcDrawableNormal);
+                                int backgroundColorNormal,
+                                int backgroundColorPressed,
+                                int backgroundColorSelected,
+                                int backgroundColorChecked,
+                                int backgroundColorUnable,
+
+                                int backgroundDrawableNormal,
+                                int backgroundDrawablePressed,
+                                int backgroundDrawableSelected,
+                                int backgroundDrawableChecked,
+                                int backgroundDrawableUnable,
+
+                                int gradientOrientationNormal,
+                                int gradientOrientationPressed,
+                                int gradientOrientationSelected,
+                                int gradientOrientationChecked,
+                                int gradientOrientationUnable,
+
+                                int gradientCenterXNormal,
+                                int gradientCenterXPressed,
+                                int gradientCenterXSelected,
+                                int gradientCenterXChecked,
+                                int gradientCenterXUnable,
+
+                                int gradientCenterYNormal,
+                                int gradientCenterYPressed,
+                                int gradientCenterYSelected,
+                                int gradientCenterYChecked,
+                                int gradientCenterYUnable,
+
+                                int gradientStartColorNormal,
+                                int gradientStartColorPressed,
+                                int gradientStartColorSelected,
+                                int gradientStartColorChecked,
+                                int gradientStartColorUnable,
+
+                                int gradientCenterColorNormal,
+                                int gradientCenterColorPressed,
+                                int gradientCenterColorSelected,
+                                int gradientCenterColorChecked,
+                                int gradientCenterColorUnable,
+
+                                int gradientEndColorNormal,
+                                int gradientEndColorPressed,
+                                int gradientEndColorSelected,
+                                int gradientEndColorChecked,
+                                int gradientEndColorUnable,
+
+                                int gradientRadiusNormal,
+                                int gradientRadiusPressed,
+                                int gradientRadiusSelected,
+                                int gradientRadiusChecked,
+                                int gradientRadiusUnable,
+
+                                int gradientTypeNormal,
+                                int gradientTypePressed,
+                                int gradientTypeSelected,
+                                int gradientTypeChecked,
+                                int gradientTypeUnable,
+
+                                int cornerRadiusNormal,
+                                int cornerRadiusPressed,
+                                int cornerRadiusSelected,
+                                int cornerRadiusChecked,
+                                int cornerRadiusUnable,
+
+                                int cornerRadiusTopLeftNormal,
+                                int cornerRadiusTopLeftPressed,
+                                int cornerRadiusTopLeftSelected,
+                                int cornerRadiusTopLeftChecked,
+                                int cornerRadiusTopLeftUnable,
+
+                                int cornerRadiusTopRightNormal,
+                                int cornerRadiusTopRightPressed,
+                                int cornerRadiusTopRightSelected,
+                                int cornerRadiusTopRightChecked,
+                                int cornerRadiusTopRightUnable,
+
+                                int cornerRadiusBottomLeftNormal,
+                                int cornerRadiusBottomLeftPressed,
+                                int cornerRadiusBottomLeftSelected,
+                                int cornerRadiusBottomLeftChecked,
+                                int cornerRadiusBottomLeftUnable,
+
+                                int cornerRadiusBottomRightNormal,
+                                int cornerRadiusBottomRightPressed,
+                                int cornerRadiusBottomRightSelected,
+                                int cornerRadiusBottomRightChecked,
+                                int cornerRadiusBottomRightUnable,
+
+                                int borderWidthNormal,
+                                int borderWidthPressed,
+                                int borderWidthSelected,
+                                int borderWidthChecked,
+                                int borderWidthUnable,
+
+                                int borderDashWidthNormal,
+                                int borderDashWidthPressed,
+                                int borderDashWidthSelected,
+                                int borderDashWidthChecked,
+                                int borderDashWidthUnable,
+
+                                int borderDashGapNormal,
+                                int borderDashGapPressed,
+                                int borderDashGapSelected,
+                                int borderDashGapChecked,
+                                int borderDashGapUnable,
+
+                                int borderColorNormal,
+                                int borderColorPressed,
+                                int borderColorSelected,
+                                int borderColorChecked,
+                                int borderColorUnable) {
+
+        super(imageView,
+                typedArray,
+
+                backgroundColorNormal,
+                backgroundColorPressed,
+                backgroundColorSelected,
+                backgroundColorChecked,
+                backgroundColorUnable,
+
+                backgroundDrawableNormal,
+                backgroundDrawablePressed,
+                backgroundDrawableSelected,
+                backgroundDrawableChecked,
+                backgroundDrawableUnable,
+
+                gradientOrientationNormal,
+                gradientOrientationPressed,
+                gradientOrientationSelected,
+                gradientOrientationChecked,
+                gradientOrientationUnable,
+
+                gradientCenterXNormal,
+                gradientCenterXPressed,
+                gradientCenterXSelected,
+                gradientCenterXChecked,
+                gradientCenterXUnable,
+
+                gradientCenterYNormal,
+                gradientCenterYPressed,
+                gradientCenterYSelected,
+                gradientCenterYChecked,
+                gradientCenterYUnable,
+
+                gradientStartColorNormal,
+                gradientStartColorPressed,
+                gradientStartColorSelected,
+                gradientStartColorChecked,
+                gradientStartColorUnable,
+
+                gradientCenterColorNormal,
+                gradientCenterColorPressed,
+                gradientCenterColorSelected,
+                gradientCenterColorChecked,
+                gradientCenterColorUnable,
+
+                gradientEndColorNormal,
+                gradientEndColorPressed,
+                gradientEndColorSelected,
+                gradientEndColorChecked,
+                gradientEndColorUnable,
+
+                gradientRadiusNormal,
+                gradientRadiusPressed,
+                gradientRadiusSelected,
+                gradientRadiusChecked,
+                gradientRadiusUnable,
+
+                gradientTypeNormal,
+                gradientTypePressed,
+                gradientTypeSelected,
+                gradientTypeChecked,
+                gradientTypeUnable,
+
+                cornerRadiusNormal,
+                cornerRadiusPressed,
+                cornerRadiusSelected,
+                cornerRadiusChecked,
+                cornerRadiusUnable,
+
+                cornerRadiusTopLeftNormal,
+                cornerRadiusTopLeftPressed,
+                cornerRadiusTopLeftSelected,
+                cornerRadiusTopLeftChecked,
+                cornerRadiusTopLeftUnable,
+
+                cornerRadiusTopRightNormal,
+                cornerRadiusTopRightPressed,
+                cornerRadiusTopRightSelected,
+                cornerRadiusTopRightChecked,
+                cornerRadiusTopRightUnable,
+
+                cornerRadiusBottomLeftNormal,
+                cornerRadiusBottomLeftPressed,
+                cornerRadiusBottomLeftSelected,
+                cornerRadiusBottomLeftChecked,
+                cornerRadiusBottomLeftUnable,
+
+                cornerRadiusBottomRightNormal,
+                cornerRadiusBottomRightPressed,
+                cornerRadiusBottomRightSelected,
+                cornerRadiusBottomRightChecked,
+                cornerRadiusBottomRightUnable,
+
+                borderWidthNormal,
+                borderWidthPressed,
+                borderWidthSelected,
+                borderWidthChecked,
+                borderWidthUnable,
+
+                borderDashWidthNormal,
+                borderDashWidthPressed,
+                borderDashWidthSelected,
+                borderDashWidthChecked,
+                borderDashWidthUnable,
+
+                borderDashGapNormal,
+                borderDashGapPressed,
+                borderDashGapSelected,
+                borderDashGapChecked,
+                borderDashGapUnable,
+
+                borderColorNormal,
+                borderColorPressed,
+                borderColorSelected,
+                borderColorChecked,
+                borderColorUnable);
+
+        mSrcNormal = typedArray.getDrawable(srcNormal);
+        mSrcPressed = typedArray.getDrawable(srcPressed);
+        mSrcSelected = typedArray.getDrawable(srcSelected);
+        mSrcChecked = typedArray.getDrawable(srcChecked);
+        mSrcUnable = typedArray.getDrawable(srcUnable);
+
+        if (mSrcPressed == null && mSrcNormal != null) {
+            mSrcPressed = typedArray.getDrawable(srcNormal);
         }
 
-        if (mSrcDrawableSelected == null && mSrcDrawableNormal != null) {
-            mSrcDrawableSelected = typedArray.getDrawable(srcDrawableNormal);
+        if (mSrcSelected == null && mSrcNormal != null) {
+            mSrcSelected = typedArray.getDrawable(srcNormal);
         }
 
-        if (mSrcDrawableChecked == null && mSrcDrawableNormal != null) {
-            mSrcDrawableChecked = typedArray.getDrawable(srcDrawableNormal);
+        if (mSrcChecked == null && mSrcNormal != null) {
+            mSrcChecked = typedArray.getDrawable(srcNormal);
         }
 
-        if (mSrcDrawableUnable == null && mSrcDrawableNormal != null) {
-            mSrcDrawableUnable = typedArray.getDrawable(srcDrawableNormal);
+        if (mSrcUnable == null && mSrcNormal != null) {
+            mSrcUnable = typedArray.getDrawable(srcNormal);
         }
 
         updateSrcDrawable();
@@ -105,82 +289,85 @@ public class ColorImageViewHelper extends ColorViewHelper<ImageView> {
     private void updateSrcDrawable() {
         StateListDrawable stateListDrawable = new StateListDrawable();
 
-        if (mSrcDrawableUnable != null) {
-            stateListDrawable.addState(mStateArray[0], mSrcDrawableUnable);
+        if (mSrcUnable != null) {
+            stateListDrawable.addState(Constant.STATE_0_UNABLE, mSrcUnable);
         }
 
-        if (mSrcDrawablePressed != null) {
-            stateListDrawable.addState(mStateArray[1], mSrcDrawablePressed);
+        if (mSrcPressed != null) {
+            stateListDrawable.addState(Constant.STATE_1_PRESSED, mSrcPressed);
         }
 
-        if (mSrcDrawableSelected != null) {
-            stateListDrawable.addState(mStateArray[2], mSrcDrawableSelected);
+        if (mSrcSelected != null) {
+            stateListDrawable.addState(Constant.STATE_2_SELECTED, mSrcSelected);
         }
 
-        if (mSrcDrawableChecked != null) {
-            stateListDrawable.addState(mStateArray[3], mSrcDrawableChecked);
+        if (mSrcChecked != null) {
+            stateListDrawable.addState(Constant.STATE_3_CHECKED, mSrcChecked);
         }
 
-        if (mSrcDrawableNormal != null) {
-            stateListDrawable.addState(mStateArray[4], mSrcDrawableNormal);
+        if (mSrcNormal != null) {
+            stateListDrawable.addState(Constant.STATE_4_ENABLED, mSrcNormal);
         }
 
-        if (mSrcDrawableNormal != null) {
-            stateListDrawable.addState(mStateArray[5], mSrcDrawableNormal);
+        if (mSrcNormal != null) {
+            stateListDrawable.addState(Constant.STATE_5_NONE, mSrcNormal);
         }
 
         mView.setImageDrawable(stateListDrawable);
     }
 
-    // SET SrcDrawable
-    public void setSrcDrawable(Drawable srcDrawable, boolean includeUnableState) {
-        if (srcDrawable == null) {
-            return;
-        }
-
-        mSrcDrawableNormal = srcDrawable;
-        mSrcDrawablePressed = srcDrawable;
-        mSrcDrawableSelected = srcDrawable;
-        mSrcDrawableChecked = srcDrawable;
-
-        if (includeUnableState) {
-            mSrcDrawableUnable = srcDrawable;
-        }
-
-        updateSrcDrawable();
+    // GET SET
+    public Drawable getSrcNormal() {
+        return mSrcNormal;
     }
 
-    public void setSrcDrawableNormal(Drawable srcDrawableNormal) {
-        if (srcDrawableNormal != null) {
-            mSrcDrawableNormal = srcDrawableNormal;
+    public void setSrcNormal(Drawable srcNormal) {
+        if (mSrcNormal != srcNormal) {
+            mSrcNormal = srcNormal;
             updateSrcDrawable();
         }
     }
 
-    public void setSrcDrawablePressed(Drawable srcDrawablePressed) {
-        if (srcDrawablePressed != null) {
-            mSrcDrawablePressed = srcDrawablePressed;
+    public Drawable getSrcPressed() {
+        return mSrcPressed;
+    }
+
+    public void setSrcPressed(Drawable srcPressed) {
+        if (mSrcPressed != srcPressed) {
+            mSrcPressed = srcPressed;
             updateSrcDrawable();
         }
     }
 
-    public void setSrcDrawableSelected(Drawable srcDrawableSelected) {
-        if (srcDrawableSelected != null) {
-            mSrcDrawableSelected = srcDrawableSelected;
+    public Drawable getSrcSelected() {
+        return mSrcSelected;
+    }
+
+    public void setSrcSelected(Drawable srcSelected) {
+        if (mSrcSelected != srcSelected) {
+            mSrcSelected = srcSelected;
             updateSrcDrawable();
         }
     }
 
-    public void setSrcDrawableChecked(Drawable srcDrawableChecked) {
-        if (srcDrawableChecked != null) {
-            mSrcDrawableChecked = srcDrawableChecked;
+    public Drawable getSrcChecked() {
+        return mSrcChecked;
+    }
+
+    public void setSrcChecked(Drawable srcChecked) {
+        if (mSrcChecked != srcChecked) {
+            mSrcChecked = srcChecked;
             updateSrcDrawable();
         }
     }
 
-    public void setSrcDrawableUnable(Drawable srcDrawableUnable) {
-        if (srcDrawableUnable != null) {
-            mSrcDrawableUnable = srcDrawableUnable;
+    public Drawable getSrcUnable() {
+        return mSrcUnable;
+    }
+
+    public void setSrcUnable(Drawable srcUnable) {
+        if (mSrcUnable != srcUnable) {
+            mSrcUnable = srcUnable;
             updateSrcDrawable();
         }
     }
